backend/controllers/IA/requirements.txt
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0
keras==2.6.0
matplotlib==3.4.3
seaborn==0.11.2
scipy==1.7.1
jupyter==1.0.0
notebook==6.4.3
python-dotenv==1.0.1
pymupdf==1.24.11
pillow==9.4.0